move config to apis
