+++ pkg/scheduler/apis/default_plugins.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package config
+package apis
 
 import (
 	"k8s.io/apimachinery/pkg/util/sets"
