+++ pkg/scheduler/apis/defaults.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package config
+package apis
 
 import (
 	"github.com/clusternet/clusternet/pkg/apis/scheduler"
